Account create, update, delete done

+++ components/ui/data-table.tsx
@@ -3,6 +3,7 @@ import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Trash } from "lucide-react"
 import * as React from "react"
+import { useConfirm } from "@/hooks/use-confirm"
 import {
     ColumnDef,
     ColumnFiltersState,
@@ -43,6 +44,10 @@ export function DataTable<TData, TValue>({
     onDelete,
     disabled,
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmationDialog, confirm] = useConfirm(
+        "Are you sure?",
+        "You are about to perform a bulk delete."
+    );
     const [rowSelection, setRowSelection] = React.useState({})
 
     const [sorting, setSorting] = React.useState<SortingState>([])
@@ -71,7 +76,9 @@ export function DataTable<TData, TValue>({
     });
 
     return (
+
         <div>
+            <ConfirmationDialog />
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey}...`}
@@ -87,9 +94,16 @@ export function DataTable<TData, TValue>({
                         size="sm"
                         variant="outline"
                         className="ml-auto font-normal text-xs"
-                        onClick={() => {
-                            onDelete(table.getFilteredSelectedRowModel().rows)
-                            table.resetRowSelection();
+                        onClick={async () => {
+                            const ok = await confirm();
+                            if (ok) {
+
+
+
+                                onDelete(table.getFilteredSelectedRowModel().rows)
+                                table.resetRowSelection();
+
+                            }
                         }}>
                         <Trash className="size-4 mr-2" />
                         Delete ({table.getFilteredSelectedRowModel().rows.length})
